# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Tool Tracker - Gibikey Studio
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  profileImage?: string | null;
}

export interface Log {
  id: number;
  text: string;
  date: string;
  completed: boolean;
}

export interface LogInput {
  text: string;
}

export interface LogUpdateInput {
  /** @nullable */
  text?: string | null;
  /** @nullable */
  completed?: boolean | null;
}

export type ToolStatus = (typeof ToolStatus)[keyof typeof ToolStatus];

export const ToolStatus = {
  Pending: "Pending",
  Deploy: "Deploy",
  Publish: "Publish",
} as const;

export interface Tool {
  id: number;
  name: string;
  description: string;
  url: string;
  status: ToolStatus;
  createdWith: string;
  createdByAccount: string;
  deployWith: string;
  deployByAccount: string;
  version: string;
  releaseDate: string;
  createdAt: string;
  logs: Log[];
}

export type ToolInputStatus =
  (typeof ToolInputStatus)[keyof typeof ToolInputStatus];

export const ToolInputStatus = {
  Pending: "Pending",
  Deploy: "Deploy",
  Publish: "Publish",
} as const;

export interface ToolInput {
  name: string;
  description?: string;
  url?: string;
  status?: ToolInputStatus;
  createdWith?: string;
  createdByAccount?: string;
  deployWith?: string;
  deployByAccount?: string;
  version?: string;
  releaseDate?: string;
}

export type PlanningCategory =
  (typeof PlanningCategory)[keyof typeof PlanningCategory];

export const PlanningCategory = {
  Ide: "Ide",
  Untuk_Dijual: "Untuk Dijual",
  Portofolio: "Portofolio",
  Internal: "Internal",
  Abandon: "Abandon",
} as const;

export interface Planning {
  id: number;
  name: string;
  description: string;
  url: string;
  category: PlanningCategory;
  price: string;
  target: string;
  createdAt: string;
  logs: Log[];
}

export type PlanningInputCategory =
  (typeof PlanningInputCategory)[keyof typeof PlanningInputCategory];

export const PlanningInputCategory = {
  Ide: "Ide",
  Untuk_Dijual: "Untuk Dijual",
  Portofolio: "Portofolio",
  Internal: "Internal",
  Abandon: "Abandon",
} as const;

export interface PlanningInput {
  name: string;
  description?: string;
  url?: string;
  category?: PlanningInputCategory;
  price?: string;
  target?: string;
}

export interface AccountInput {
  email: string;
}

export interface Stats {
  totalTools: number;
  publishedTools: number;
  totalPlannings: number;
  potentialRevenue: number;
}