# lib/api-zod/src/generated/types/accountInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Tool Tracker - Gibikey Studio
 * OpenAPI spec version: 0.1.0
 */

export interface AccountInput {
  email: string;
}