# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Tool Tracker - Gibikey Studio
 * OpenAPI spec version: 0.1.0
 */

export interface AuthUser {
  id: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  profileImage?: string | null;
}